requirements.txt
gradio>=4.0.0
google-genai
python-dotenv
pyaudio
opencv-python
Pillow
mss
numpy
asyncio

